Reorganize such that pure C# code is contained in a common file, and there are engine-specific wrappers for the UDP functionality.
